requirements.txt
asgiref==3.7.2
black==24.2.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.4
defusedxml==0.8.0rc2
distlib==0.3.8
Django==4.2
django-cors-headers
django-oauth-toolkit==2.3.0
django-rest-framework-social-oauth2==1.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-jwt
djangorestframework-simplejwt
djoser==2.1.0
django-filter
filelock==3.13.1
identify==2.5.35
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
jwcrypto==1.5.4
MarkupSafe==2.1.5
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.6.2
psycopg2-binary==2.9.3
pycparser==2.21
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.1
